`@0x:contract-integrations` Move caches to a new file
